Meowbot/requirements.txt
Flask>=1.1
flasgger >= 0.9.7.1
flask-swagger-ui >= 4.11.1
pymongo >= 4.3.3
line-bot-sdk >= 2.4.2
openai >= 0.27.8